move HandleProwJob from Sub to Gangway
Now both Sub and Gangway use the same HandleProwJob function to convert
a CreateJobExecutionRequest to a Prow Job CR. All integration tests are
passing.

We had to move out the HandleProwJob function and other smaller
functions around it into Gangway to avoid cyclic imports.
